+++ Source/Scripts/PackageManagement/Install-PackageManagement.ps1
@@ -6,6 +6,10 @@
 )        
 
 $cmdletsPaths="$PSScriptRoot\..\..\Cmdlets"
+
+. "$cmdletsPaths\Helpers\Write-MyInvocation.ps1"
+Write-MyInvocation -Invocation $MyInvocation
+
 . "$cmdletsPaths\Helpers\Invoke-CommandWrap.ps1"
 
 $packageManagementScriptBlock={