+++ Source/Scripts/xISHServer/Upload-ISHServerPrerequisites.ps1
@@ -7,12 +7,17 @@
     [ValidateSet("12","13")]
     [string]$ISHServerVersion
 )    
-. $PSScriptRoot\..\..\Cmdlets\Helpers\Invoke-CommandWrap.ps1
+$cmdletsPaths="$PSScriptRoot\..\..\Cmdlets"
+
+. "$cmdletsPaths\Helpers\Write-MyInvocation.ps1"
+Write-MyInvocation -Invocation $MyInvocation
+
+. "$cmdletsPaths\Helpers\Invoke-CommandWrap.ps1"
 
 if($Computer)
 {
-    . $PSScriptRoot\..\..\Cmdlets\Helpers\Add-ModuleFromRemote.ps1
-    . $PSScriptRoot\..\..\Cmdlets\Helpers\Remove-ModuleFromRemote.ps1
+    . $cmdletsPaths\Helpers\Add-ModuleFromRemote.ps1
+    . $cmdletsPaths\Helpers\Remove-ModuleFromRemote.ps1
 }
 
 try